Update Models and UpdateServingTime Funciton Update

+++ Public/Agent.cs
@@ -1,11 +1,9 @@
 using Newtonsoft.Json;
 
-
 namespace LAVI.QueueManager
 {
     public class Agent
     {
-
         [JsonProperty("agentId")]
         public string agentId { get; set; }
 
@@ -16,13 +14,12 @@ namespace LAVI.QueueManager
         public string[] queueIds { get; set; }
 
         [JsonProperty("busyTillTimeInMilliseconds")]
-        public int busyTillTimeInMilliseconds { get; set; }
+        public double busyTillTimeInMilliseconds { get; set; }
 
         [JsonProperty("customersInServing")]
         public KioskRequest[] customersInServing { get; set; }
         [JsonProperty("customerToBeServed")]
         public KioskRequest[] customerToBeServed { get; set; }
-
     }
 }
